e2e: factor proxy test out of regression test
