+++ .github/workflows/e2e_manual.yml
@@ -17,6 +17,7 @@ on:
           - openssl
           - peerrecovery
           - policy
+          - proxy
           - regression
           - servicemesh
           - vault
